client/requirements.txt
websocket-client==1.3.2